memory barriers
git-svn-id: https://groke.otago.ac.nz/ant/ant@784 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/stats_memory_index.c
@@ -66,7 +66,7 @@ if (type & STAT_SUMMARY)
 if (type & STAT_MEMORY)
 	{
 	printf("\nINTERNAL MEMORY BREAKDOWN\n-------------------------\n");
-	printf("Hash nodes total     :%10d nodes\n", ANT_memory_index::HASH_TABLE_SIZE);
+	printf("Hash nodes total     :%10d nodes\n", (int)ANT_memory_index::HASH_TABLE_SIZE);
 	printf("Hash nodes used      :%10ld nodes\n", hash_nodes);
 	printf("Hash nodes used      :%10ld bytes\n", (long)(hash_nodes * sizeof(ANT_memory_index_hash_node)));
 	printf("HASH utilisation     :%10.2f%%\n", ((double)hash_nodes / (double)ANT_memory_index::HASH_TABLE_SIZE) * 100);
